all tests refactored
