+++ tests/PageObjects/AfterRegisterPage.py
@@ -1,0 +1,13 @@
+import time
+import selenium
+from selenium.webdriver import ActionChains
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import Select
+from selenium.common.exceptions import TimeoutException
+from selenium.webdriver.support.wait import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+from PageObjects.BasePage import BasePage
+
+
+class AfterRegisterPage(BasePage):
+    SUCCESS_NOTE = (By.CSS_SELECTOR, '#content > h1')
