+++ tests/test_iphone_page.py
@@ -1,29 +1,28 @@
-from logging import addLevelName
-import pytest
-from conftest import driver, url
-from exception_wrappers import *
-from PageObjects import *
+import time
+import selenium
+from PageObjects.MainPage import MainPage
+from PageObjects.ProductPage import ProductPage
 
 
 def test_product_button_in_main(driver, url):
     driver.get(url=url)
     iphone_ref = MainPage(driver).element(MainPage.IPHONE)
-    MainPage.click(iphone_ref)
+    MainPage(driver).click(iphone_ref)
     product_name = ProductPage(driver).get_product_name()
-    assert product_name.text == "iPhone"
+    assert product_name == "iPhone"
 
 
 def test_add_to_cart(driver, url):
     driver.get(url=url+'/iphone')
     ProductPage(driver).add_to_cart()
-    alert = ProductPage.element(AlertElement.CART)
+    alert = ProductPage(driver).element(ProductPage.ALERT_ADD_TO_CART_SUCCESS)
     assert "You have added" in alert.text
 
 
 def test_wish_list(driver, url):
     driver.get(url=url+"/iphone")
     ProductPage(driver).add_to_wish_list()
-    alert = ProductPage.element(AlertElement.WISH)
+    alert = ProductPage(driver).element(ProductPage.ALERT_ADD_TO_WISH_UNSUCCESS)
     assert "You must" in alert.text
 
 
@@ -38,3 +37,31 @@ def test_related_products(driver, url):
 
     product_image =ProductPage(driver).element_in_element(related_product, ProductPage.RELATED_PRODUCT_IMAGE)
     assert product_image.get_attribute('src').endswith('.jpg')
+
+
+# from selenium import webdriver
+# driver1 = webdriver.Chrome(executable_path=r'C:\Users\marisarze\Downloads\browsers\chromedriver.exe')
+# url1 = r"http://192.168.0.102:8081"
+# test_product_button_in_main(driver1, url1)
+
+# from selenium import webdriver
+# driver2 = webdriver.Chrome(executable_path=r'C:\Users\marisarze\Downloads\browsers\chromedriver.exe')
+# url2 = r"http://192.168.0.102:8081"
+# test_add_to_cart(driver2, url2)
+
+# from selenium import webdriver
+# driver4 = webdriver.Chrome(executable_path=r'C:\Users\marisarze\Downloads\browsers\chromedriver.exe')
+# url4 = r"http://192.168.0.102:8081"
+# test_wish_list(driver4, url4)
+
+# from selenium import webdriver
+# driver3 = webdriver.Chrome(executable_path=r'C:\Users\marisarze\Downloads\browsers\chromedriver.exe')
+# url3 = r"http://192.168.0.102:8081"
+# test_related_products(driver3, url3)
+
+
+
+# from selenium import webdriver
+# driver5 = webdriver.Chrome(executable_path=r'C:\Users\marisarze\Downloads\browsers\chromedriver.exe')
+# url5 = r"http://192.168.0.102:8081"
+# test_basket(driver5,url5)
